правки в update

+++ services/tile.py
@@ -95,12 +95,15 @@ async def add_tile(
 
 async def delete_tile(manager, file_manager, uow_class=UnitOfWork, **filters):
     async with uow_class(manager) as uow:
-        tiles = await manager.read(
+        tiles_to_delete = await manager.read(
             Tile, loaded=["images"], session=uow.session, **filters
         )
+        if not tiles_to_delete:
+            return []
+
         del_res = await manager.delete(Tile, session=uow.session, **filters)
-        for tile in tiles:
-            images_paths = tile.get("images_paths", [])
+        for tile in tiles_to_delete:
+            images_paths = tile.images_paths
             for image in images_paths:
                 await file_manager.delete_product(image)
         return del_res
@@ -137,9 +140,20 @@ def set_values_from_db(values: dict, key: str, value_from_db):
         values[key] = value_from_db
 
 
-async def create_composite(manager, article: int, values, columns, *to_join):
-    model = (await manager.read(Tile, loaded=list(to_join), id=article))[0]
-    for k, v in model.items():
+# async def create_composite(manager, article: int, values, columns, *to_join):
+#     model = await manager.read_one(Tile, loaded=list(to_join), id=article)
+#     if not model:
+#         return
+#     for k, v in model.items():
+#         if k in columns:
+#             k = map_tile_param_to_model_param(k)
+#             set_values_from_db(values, k, v)
+async def create_composite(manager, article: int, values: dict, columns: tuple, *to_join):
+    tile = await manager.read_one(Tile, loaded=list(to_join), id=article)
+    if not tile:
+        return
+
+    for k, v in tile.to_dict().items():
         if k in columns:
             k = map_tile_param_to_model_param(k)
             set_values_from_db(values, k, v)
@@ -194,36 +208,81 @@ def map_param_to_domain_model(param_name: str):
     return mapper[param_name]
 
 
+# async def update_tile(
+#     manager,
+#     article: int,
+#     uow_class=UnitOfWork,
+#     name: str | None = None,
+#     size: dict | None = None,
+#     color: dict | None = None,
+#     producer_name: str | None = None,
+#     box: dict | None = None,
+#     boxes_count: int | None = None,
+#     category_name: str | None = None,
+#     surface_name: str | None = None,
+# ):
+#     params = {
+#         k: v
+#         for k, v in locals().items()
+#         if v is not None and k not in {"manager", "article", "uow_class"}
+#     }
+#     to_update = {}
+#     async with uow_class(manager) as uow:
+#         for k, v in params.items():
+#             domain_model = map_param_to_domain_model(k)
+#             if domain_model is Tile:
+#                 to_update.update({k: v})
+#                 continue
+#             updated_in_model = dict_for_update_model(k, v)
+#             updated_fields_in_tile = await create_new_model(
+#                 manager, article, domain_model, uow.session, **updated_in_model
+#             )
+#             to_update.update(updated_fields_in_tile)
+#         await manager.update(
+#             Tile, session=uow.session, filters=dict(id=article), **to_update
+#         )
+
+
 async def update_tile(
     manager,
     article: int,
+    size: SizeUpdate,
+    color: ColorUpdate,
+    producer_name: ProducerUpdate,
+    box: BoxUpdate,
+    category_name: CategoryUpdate,
+    surface_name: SurfaceUpdate,
     uow_class=UnitOfWork,
     name: str | None = None,
-    size: dict | None = None,
-    color: dict | None = None,
-    producer_name: str | None = None,
-    box: dict | None = None,
     boxes_count: int | None = None,
-    category_name: str | None = None,
-    surface_name: str | None = None,
 ):
     params = {
-        k: v
-        for k, v in locals().items()
-        if v is not None and k not in {"manager", "article", "uow_class"}
+        k: v for k, v in locals().items()
+        if v is not None and k not in {"manager", "article", "uow_class", "name", "boxes_count"}
     }
+
     to_update = {}
+    if name is not None: to_update["name"] = name
+    if boxes_count is not None: to_update["boxes_count"] = boxes_count
+
     async with uow_class(manager) as uow:
-        for k, v in params.items():
-            domain_model = map_param_to_domain_model(k)
-            if domain_model is Tile:
-                to_update.update({k: v})
+        commands_needing_read = [cmd for cmd in params.values() if cmd.need_read()]
+
+        if commands_needing_read:
+            needed_relations = [cmd.relation_name for cmd in commands_needing_read]
+            record = await manager.read_one(Tile, id=article, loaded=needed_relations, session=uow.session)
+            tile_flat_dict = record.to_dict() if record else {}
+
+            for cmd in commands_needing_read:
+                cmd.set_fields(tile_flat_dict)
+
+        for param in params.values():
+            if not param.need_update():
                 continue
-            updated_in_model = dict_for_update_model(k, v)
-            updated_fields_in_tile = await create_new_model(
-                manager, article, domain_model, uow.session, **updated_in_model
-            )
-            to_update.update(updated_fields_in_tile)
+            new_value = await manager.create(param.domain_model(), session=uow.session)
+            for target_field, source_attr in param.output_map.items():
+                to_update[target_field] = getattr(new_value, source_attr)
+
         await manager.update(
             Tile, session=uow.session, filters=dict(id=article), **to_update
         )
